feat(Factory): remove targetOptions from required options
If the factory does not need to save the data of its objects
you may don't set targetOptions. In this case, encrypt function
will return array of uuid.

But if you need object, you may set targetOptions. In this case,
encrypt function will return array of objects with uuid and
options which you setted.